Use correct join game method
The other bug that PLAYING is not displayed or games not removed is fixed in a different PR, that reworks code parsing game changes from the server. See PR #1261...
Fixes #1259
